refactor: model 타입 리팩토링 #17
