+++ app/src/main/java/com/ikseong/ucompass/ui/model/Direction.kt
@@ -1,0 +1,12 @@
+package com.ikseong.ucompass.ui.model
+
+enum class Direction(val rotation: Float) {
+    N(0f),
+    NE(45f),
+    E(90f),
+    SE(135f),
+    S(180f),
+    SW(225f),
+    W(270f),
+    NW(315f)
+}
